+++ tests/integration/test_models/test_full_battery_models/test_lead_acid/test_loqs.py
@@ -47,7 +47,7 @@ class TestLOQS(unittest.TestCase):
         model = pybamm.lead_acid.LOQS()
         parameter_values = model.default_parameter_values
         parameter_values.update(
-            {"Current function": 0}
+            {"Current function": "[zero]"}
         )
         modeltest = tests.StandardModelTest(model, parameter_values=parameter_values)
         modeltest.test_all()