+++ tests/integration/test_models/test_full_battery_models/test_lead_acid/test_side_reactions/test_composite_side_reactions.py
@@ -35,7 +35,7 @@ class TestLeadAcidCompositeSideReactions(unittest.TestCase):
         options = {"side reactions": ["oxygen"], "surface form": "differential"}
         model = pybamm.lead_acid.Composite(options)
         parameter_values = model.default_parameter_values
-        parameter_values.update({"Current function": 0})
+        parameter_values.update({"Current function": "[zero]"})
         modeltest = tests.StandardModelTest(model, parameter_values=parameter_values)
         modeltest.test_all(skip_output_tests=True)